[78] Move detail feature
