+++ app/src/test/java/co/nimblehq/compose/crypto/ui/screens/detail/DetailViewModelTest.kt
@@ -3,9 +3,10 @@ package co.nimblehq.compose.crypto.ui.screens.detail
 import app.cash.turbine.test
 import co.nimblehq.compose.crypto.domain.usecase.GetCoinDetailUseCase
 import co.nimblehq.compose.crypto.domain.usecase.GetCoinPricesUseCase
+import co.nimblehq.compose.crypto.feature.detail.DetailViewModel
+import co.nimblehq.compose.crypto.feature.util.toUiModel
 import co.nimblehq.compose.crypto.test.MockUtil
 import co.nimblehq.compose.crypto.ui.screens.BaseViewModelTest
-import co.nimblehq.compose.crypto.ui.uimodel.toUiModel
 import io.kotest.matchers.shouldBe
 import io.mockk.every
 import io.mockk.mockk
