+++ feature/home/src/main/java/co/nimblehq/compose/crypto/feature/util/ExtUtil.kt
@@ -1,7 +1,11 @@
 package co.nimblehq.compose.crypto.feature.util
 
+import co.nimblehq.compose.crypto.core.uimodel.CoinDetailUiModel
 import co.nimblehq.compose.crypto.core.uimodel.CoinItemUiModel
+import co.nimblehq.compose.crypto.data.extension.orZero
 import co.nimblehq.compose.crypto.domain.home.model.CoinItem
+import co.nimblehq.compose.crypto.domain.model.CoinDetail
+import co.nimblehq.compose.crypto.feature.home.FIAT_CURRENCY
 
 fun CoinItem.toUiModel() = CoinItemUiModel(
     id = id,
@@ -10,4 +14,18 @@ fun CoinItem.toUiModel() = CoinItemUiModel(
     image = image,
     currentPrice = currentPrice,
     priceChangePercentage24hInCurrency = priceChangePercentage24hInCurrency
+)
+
+fun CoinDetail.toUiModel() = CoinDetailUiModel(
+    coinName = coinName,
+    image = image?.small.orEmpty(),
+    currentPrice = marketData?.currentPrice?.getValue(FIAT_CURRENCY).orZero(),
+    priceChangePercentage24hInCurrency = marketData?.priceChangePercentage24hInCurrency
+        ?.getValue(FIAT_CURRENCY).orZero(),
+    marketCap = marketData?.marketCap?.getValue(FIAT_CURRENCY).orZero(),
+    marketCapChangePercentage24h = marketData?.marketCapChangePercentage24h.orZero(),
+    ath = marketData?.ath?.getValue(FIAT_CURRENCY).orZero(),
+    athChangePercentage = marketData?.athChangePercentage?.getValue(FIAT_CURRENCY).orZero(),
+    atl = marketData?.atl?.getValue(FIAT_CURRENCY).orZero(),
+    atlChangePercentage = marketData?.atlChangePercentage?.getValue(FIAT_CURRENCY).orZero()
 )
