+++ feature/src/main/java/co/nimblehq/compose/crypto/feature/detail/SellBuyGroup.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.compose.crypto.ui.screens.detail
+package co.nimblehq.compose.crypto.feature.detail
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
